[engine]Fix issue #30 Tutorial 08 is misplaced totally out of screen in OpenGL
Fixed by not setting the transform on entities which have 'Null' matrix. This is original
code logic but makes no sense for me. Maybe some specific cases need these checks but as far as
the tutorial is concerned, I've tested in DirectX and tutorial works the same without
this code... so why to keep it? :)